refactor: rename RequiredVersion to PlatformHelperVersioning; update imports and test patching paths; consolidate PlatformHelperVersioning provider functions into domain
